Snow - Fix circular dependency.
Move the utility methods for the NASDE snow models inside the module
since that is the only location where they will be used and fixes the
circular bad design.
